requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==4.2.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cx-Oracle==8.3.0
decorator==4.4.2
Django==4.0.4
django-appconf==1.0.5
django-compressor==4.0
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-libsass==0.9
django-model-utils==4.2.0
django-rest-framework==0.1.0
django-sass-processor==1.2
django-shopping-cart==0.1
django-unicorn==0.45.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
dnspython==2.2.1
docopt==0.6.2
filelock==3.7.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
Js2Py==0.71
libsass==0.21.0
numpy==1.21.2
orjson==3.7.2
packaging==21.3
Pillow==9.1.1
PyAudio==0.2.11
pyjsparser==2.7.1
pyobjc==8.5
pyobjc-core==8.5
pyobjc-framework-Accessibility==8.5
pyobjc-framework-Accounts==8.5
pyobjc-framework-AddressBook==8.5
pyobjc-framework-AdServices==8.5
pyobjc-framework-AdSupport==8.5
pyobjc-framework-AppleScriptKit==8.5
pyobjc-framework-AppleScriptObjC==8.5
pyobjc-framework-ApplicationServices==8.5
pyobjc-framework-AppTrackingTransparency==8.5
pyobjc-framework-AudioVideoBridging==8.5
pyobjc-framework-AuthenticationServices==8.5
pyobjc-framework-AutomaticAssessmentConfiguration==8.5
pyobjc-framework-Automator==8.5
pyobjc-framework-AVFoundation==8.5
pyobjc-framework-AVKit==8.5
pyobjc-framework-BusinessChat==8.5
pyobjc-framework-CalendarStore==8.5
pyobjc-framework-CallKit==8.5
pyobjc-framework-CFNetwork==8.5
pyobjc-framework-ClassKit==8.5
pyobjc-framework-CloudKit==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-Collaboration==8.5
pyobjc-framework-ColorSync==8.5
pyobjc-framework-Contacts==8.5
pyobjc-framework-ContactsUI==8.5
pyobjc-framework-CoreAudio==8.5
pyobjc-framework-CoreAudioKit==8.5
pyobjc-framework-CoreBluetooth==8.5
pyobjc-framework-CoreData==8.5
pyobjc-framework-CoreHaptics==8.5
pyobjc-framework-CoreLocation==8.5
pyobjc-framework-CoreMedia==8.5
pyobjc-framework-CoreMediaIO==8.5
pyobjc-framework-CoreMIDI==8.5
pyobjc-framework-CoreML==8.5
pyobjc-framework-CoreMotion==8.5
pyobjc-framework-CoreServices==8.5
pyobjc-framework-CoreSpotlight==8.5
pyobjc-framework-CoreText==8.5
pyobjc-framework-CoreWLAN==8.5
pyobjc-framework-CryptoTokenKit==8.5
pyobjc-framework-DataDetection==8.5
pyobjc-framework-DeviceCheck==8.5
pyobjc-framework-DictionaryServices==8.5
pyobjc-framework-DiscRecording==8.5
pyobjc-framework-DiscRecordingUI==8.5
pyobjc-framework-DiskArbitration==8.5
pyobjc-framework-DVDPlayback==8.5
pyobjc-framework-EventKit==8.5
pyobjc-framework-ExceptionHandling==8.5
pyobjc-framework-ExecutionPolicy==8.5
pyobjc-framework-ExternalAccessory==8.5
pyobjc-framework-FileProvider==8.5
pyobjc-framework-FileProviderUI==8.5
pyobjc-framework-FinderSync==8.5
pyobjc-framework-FSEvents==8.5
pyobjc-framework-GameCenter==8.5
pyobjc-framework-GameController==8.5
pyobjc-framework-GameKit==8.5
pyobjc-framework-GameplayKit==8.5
pyobjc-framework-ImageCaptureCore==8.5
pyobjc-framework-IMServicePlugIn==8.5
pyobjc-framework-InputMethodKit==8.5
pyobjc-framework-InstallerPlugins==8.5
pyobjc-framework-InstantMessage==8.5
pyobjc-framework-Intents==8.5
pyobjc-framework-IntentsUI==8.5
pyobjc-framework-IOSurface==8.5
pyobjc-framework-iTunesLibrary==8.5
pyobjc-framework-KernelManagement==8.5
pyobjc-framework-LatentSemanticMapping==8.5
pyobjc-framework-LaunchServices==8.5
pyobjc-framework-libdispatch==8.5
pyobjc-framework-LinkPresentation==8.5
pyobjc-framework-LocalAuthentication==8.5
pyobjc-framework-LocalAuthenticationEmbeddedUI==8.5
pyobjc-framework-MailKit==8.5
pyobjc-framework-MapKit==8.5
pyobjc-framework-MediaAccessibility==8.5
pyobjc-framework-MediaLibrary==8.5
pyobjc-framework-MediaPlayer==8.5
pyobjc-framework-MediaToolbox==8.5
pyobjc-framework-Metal==8.5
pyobjc-framework-MetalKit==8.5
pyobjc-framework-MetalPerformanceShaders==8.5
pyobjc-framework-MetalPerformanceShadersGraph==8.5
pyobjc-framework-MetricKit==8.5
pyobjc-framework-MLCompute==8.5
pyobjc-framework-ModelIO==8.5
pyobjc-framework-MultipeerConnectivity==8.5
pyobjc-framework-NaturalLanguage==8.5
pyobjc-framework-NetFS==8.5
pyobjc-framework-Network==8.5
pyobjc-framework-NetworkExtension==8.5
pyobjc-framework-NotificationCenter==8.5
pyobjc-framework-OpenDirectory==8.5
pyobjc-framework-OSAKit==8.5
pyobjc-framework-OSLog==8.5
pyobjc-framework-PassKit==8.5
pyobjc-framework-PencilKit==8.5
pyobjc-framework-Photos==8.5
pyobjc-framework-PhotosUI==8.5
pyobjc-framework-PreferencePanes==8.5
pyobjc-framework-PushKit==8.5
pyobjc-framework-Quartz==8.5
pyobjc-framework-QuickLookThumbnailing==8.5
pyobjc-framework-ReplayKit==8.5
pyobjc-framework-SafariServices==8.5
pyobjc-framework-SceneKit==8.5
pyobjc-framework-ScreenCaptureKit==8.5
pyobjc-framework-ScreenSaver==8.5
pyobjc-framework-ScreenTime==8.5
pyobjc-framework-ScriptingBridge==8.5
pyobjc-framework-SearchKit==8.5
pyobjc-framework-Security==8.5
pyobjc-framework-SecurityFoundation==8.5
pyobjc-framework-SecurityInterface==8.5
pyobjc-framework-ServiceManagement==8.5
pyobjc-framework-ShazamKit==8.5
pyobjc-framework-Social==8.5
pyobjc-framework-SoundAnalysis==8.5
pyobjc-framework-Speech==8.5
pyobjc-framework-SpriteKit==8.5
pyobjc-framework-StoreKit==8.5
pyobjc-framework-SyncServices==8.5
pyobjc-framework-SystemConfiguration==8.5
pyobjc-framework-SystemExtensions==8.5
pyobjc-framework-UniformTypeIdentifiers==8.5
pyobjc-framework-UserNotifications==8.5
pyobjc-framework-UserNotificationsUI==8.5
pyobjc-framework-VideoSubscriberAccount==8.5
pyobjc-framework-VideoToolbox==8.5
pyobjc-framework-Virtualization==8.5
pyobjc-framework-Vision==8.5
pyobjc-framework-WebKit==8.5
pyparsing==3.0.9
PyPrind==2.11.3
pyserial==3.5
pySmartDL==1.3.4
pyttsx3==2.90
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
ratelim==0.1.6
rcssmin==1.1.0
requests==2.27.1
rjsmin==1.2.0
servo==1
shortuuid==1.0.9
six==1.16.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
